+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Navigation Changelog
 
+## 0.5.7
+
+**THIS UPDATE CONTAINS BREAKING CHANGES RELATED TO `NavigationNodeFactory` things**
+
+* `Code`:
+  * Replace `NavigationNodeFactory.Companion.Typed` factory into `NavigationNodeFactory.Typed`
+  * Change order of `NavigationNodeFactory.Typed` type args
+  * Add default aggregator for `NavigationNodeFactory`
+  * Add extension `changesInSubtreeFlow` for either chains and nodes
+
 ## 0.5.6
 
 * `Versions`: